use ServerGroupsAndResources to walk dynamic groups and resources
This API maintains the preferred version order, which allows us to prefer the
server-preferred version but fall back in order to other versions if necessary.
An example of where this is necessary is when querying
ImageContentSourcePolicys: at least in 4.3 these are not included in the server
preferred version, but only an older version.