Don't update unnecssarily
Operators have to decide on whether to Update() and object or not.
We can't sit in a loop and keep running update as this causes other
operators that add annotations (and watch / react to objects) to
unnecessary be busy and "fight" with us. This commit brings a part
of the v3 sync pod into the code base. This is just the bit
that ignores defaults and helps implement the above problem.