Enable boot diagnostics on RP nodes
The RP ARM template is updated to enable boot diagnostics on the VMSS. A new
storage account is also provisioned to store the bootstrap logs.

Each VMSS in a region has its own storage account.

Within the storage account, a new container is automatically created for each
node in the vmss. Only the latest bootstrap logs are stored.

This change affects the rp-production.json ARM template only.

Signed-off-by: Ivan Sim <isim@redhat.com>