Update aro deploy to support MSI auth
For deployment into fairfax we need to support MSI.  The go-autorest
supports this out of the box and we simply need to not check for the
AZURE_CLIENT_SECRET env variable for it work.  It will try each auth
method in turn until it finds one for which there are environment
variables available.  MSI is the last one and requires the smallest
configuration.