Support creating InstanceMetadata from env
In Ev2 deployments the deployer is running in Azure Container Instance
(ACI) and does _not_ have access to the instance metadata we have used
to date.  But we have the info to populate env vars.  This PR adds a
path to use env vars when AZURE_EV2 is set to non-empty value.

Includes pkg/util/env to enable testing with mocks.