Dropping ipRules parameter; it is not used in this template
It is worth noting that it is used in the rp-production.json template but leaving it in place in the rp-development.json template unused results in a syntax error that states "unable to parse template on line 1". This causes the process to hang when settings up role assignments for the shared rp.