Prevent AdminAction redeploy if VM has an ephemeral OS disk
Redeploying a VM with an ephemeral OS disk will result in a error message,
stating that the VM has an ephemeral OS disk and cannot be redeployed.

If a VM that has an ephemeral OS disk is redeployed it will never become ready,
due to the fact that the VM's OS is re-provisioned and
the cluster machine approver will not allow the "new" machine
to join the cluster again.