database: Add Backends container to CosmosDB
This container allows backends to retain state.

The immediate use for the container is to allow one and only
one backend instance to process change feed events from the
ClusterManagerConfigurations container.

It uses a kind of "leader election" pattern similar to how
the RP monitoring picks a "master" instance.  Whichever backend
instance successfully acquires the "backend lease" gets to read
from the change feed.  The other backend instances occasionally
try to acquire the lease, in case the first instance terminates
or lets the lease expire for whatever reason.

In addition, the backend container serves as a place to
persistently store the change feed's "continuation value".
This is similar to how the CosmosDB Change Feed Processor
works in CosmosDB SDKs for .NET and Java [1].

Note, this approach was preferred over adding a provisioning
state field to the ClusterManagerConfigurationDocument, which
would complicate both the frontend and backend logic.  This
approach, by contrast, is transparent to the frontend.

[1] https://docs.microsoft.com/en-us/azure/cosmos-db/sql/change-feed-processor