Add unit tests for dynamic/serviceprincipal, azureclient/environments, clusterauthorizer
ServicePrincipalChecker now accepts a function with the signature of aad.GetToken, allowing the use of the mocked GetToken function.
Add unit test for updateFailedCondition.

Add tests to azureclient/environments and clusterauthorizer that were previously in serviceprincipalchecker.
Add mock of dynamic service principal validator to prevent testing code outside of the service principal checker package.
serviceprincipal is tested by manually creating and modifying the service principal tokens.
Dynamic now contians/returns the aad.GetToken function to be used. This is to allow changing this function to a mocked aad.GetToken for testing.

Update authorizer to return a refreshable authorizer. Returning a refreshable authorizer object that is then used to get the refresable authorizer token allows for mocking the aad.GetToken calls.
Update NewValidator, NewServicePrincipalValidator, NewAzRefreshableAuthorizer to accept token client as arguement.