Run ensureResourceGroup for ARO delete
Slackthread: https://redhat-internal.slack.com/archives/C02ULBRS68M/p1676340311290279

Current Scenario:
azcli delete only ensure if ARO SP has NETWORK_CONTRIBUTOR_ROLE
on the VNET and adds it if missing, but if OWNER Permission is
missing on ARO SP for CRG, deletion fails and only AdminUpdate can fix
such a missing permission which will be tricky if the cluster is in
ProvisioningFailed state.

With this PR:
We fix any missing permission on ARO SP so that deletion works without
failure.