Fix bug in logic to preserve cluster RG tags in _putOrPatchOpenShiftCluster()
In _putOrPatchOpenShiftCluster(), we use json.Unmarshal() to take the information from the customer's request and stick it into an OpenShiftCluster struct. json.Unmarshal() merges maps rather than replacing one with the other, which is *not* the desired behavior in this case; if a customer passes a new set of tags, we want to completely replace the old set. With this in mind, ensuring that the cluster RG tags are set properly after an update to the cluster requires the addition of some non-trivial logic that essentially saves the original set of tags, allows it to be completely replaced in the initial json.Unmarshal() call by whatever the customer passed (or nil if they didn't pass any tags), and then puts the original set of tags back in place if it turns out the customer didn't pass a new set. I had initially implemented this logic incorrectly in a way that resulted in an unacceptable loss of information when patching a cluster using the admin API, so this commit fixes that bug.