chore(dynamichelper): Add warn for assumption
Added a new check for the number of GVKs that are returned because
there was an implicit assumption in the code that the kruntime.Object
passed into the function would always be fully formed therefore only
ever returning the one GVK. There's no validation of this assumption
therefore it's possible that more than one GVK is returned which may
cause unintended side-effects. This change adds a warning which may help
a future developer should there be a bug experienced because of this
assumption. Validating this assumption at this point may cause breakages
to other parts of the repository, so a warning suffices for now.