Add ResourceTags from cluster doc to resources deployed by ARM templates during cluster installation
The tagging policy assignment and its modify effect can take some time to take effect. When I was testing to see what happens when the tagging policy is applied in a subscription where resource creation is denied for resources without a particular tag, I found that the modify effect did not come into play soon enough to override the deny effect from the other policy, and so the storage template failed to deploy its resources and the cluster installation failed.

If we add `ResourceTags` to the resources in each of these ARM templates "manually" to make up for the tagging policy assignment's latency, I found experimentally that it provides enough time for the tagging policy assignment to come into effect such that cluster installation succeeds.

For the case of the integrated installer, there is plenty of time (> 5 minutes) between when the RP runs `ensureTaggingPolicy()` and the integrated installer starts trying to provision more Azure resources (eg. worker VMs).

For the case of Hive installations, I did 10 cluster installations and did not encounter any failures. I found that there was about 2-3 minutes between when the RP runs `ensureTaggingPolicy()` and the OCP installer on Hive starts trying to provision more Azure resources.