Add more robust error checking
The generic DeploymentFailed error returned by a failed ARM template deployment is a wrapper around the actual list of resource-level errors that caused the deployment failure. We need to check inside of it for known errors.

Note that the Details attribute is a list that might include more than one error, but this code only checks the first one. If there are multiple errors and the first one isn't familiar, it's probably best to just retry.