Improve error check on failed RP deployment (#2819)
* Revert "Revert "now recreates the gateway vmss if there is a change in the probe (#2784)""

This reverts commit 5b0ad2b65354bc0fda450cb1cfdc27d410166a09.

* Add more robust error checking

The generic DeploymentFailed error returned by a failed ARM template deployment is a wrapper around the actual list of resource-level errors that caused the deployment failure. We need to check inside of it for known errors.

Note that the Details attribute is a list that might include more than one error, but this code only checks the first one. If there are multiple errors and the first one isn't familiar, it's probably best to just retry.

* Reorganized error checking to be easier to read

* Add nil checks for VM health status

* Add unit tests and fix to error logging for failed deployments

* Rework condition for readability

* Remove VMSS probe stuff to consider adding in a separate PR

* Simplify logic by casting nested errors to Azure ServiceError type

* Move check of nested ServiceErrors to a helper function

* Fix lint errors

* Fix typo

* Remove nesting for simplicity