Increase gateway VMSS disk storage
This PR increaseses the OS disk to 256GB, from the default value of 64GB for both the rpVMSS and gatewayVMSS.
The root filesystem is grown to a smaller percentage now because 50% of the disk would mostly go unused.

This is needed because the gateway container writes directly to the root of it's filesystem, which is mounted in /var/containers.
The gateway log is not currently rotated, so an increase in logging can and will result in a full disk halting the VM.