Add gateway log rotatation
The gateway container log is mounted in /var/log/aro-gateway to allow logrotate access to the file.
Logs are rotated based on size, not time. Logs are rotated when they reach 20G in size, and a maximum of 100G total.
This is to allow enough space for other logs and system services in /var.

copytruncate option prevents the log file from being recreated after being deleted, which would create a condition where the log watcher/shipper thinks they are new.
copytruncate truncates the log file, leaving it's modification timestamp the same.