Remove erroneous HTTP header checks from AppLens client tests
The `pipelineVerifier` policy used to check the headers comes ahead of
the `httpHeaderPolicy`, which is what sets the request headers, so
attempting to check the headers in this way was logically incorrect from
the time the code was written. However, it just so happened to work
because the headers are stored as a `map` - a reference type. The
`httpHeaderPolicy` was modifying the same underlying `map` that the
`pipelineVerifier` stores, so the `pipelineVerifier` was being updated
with the headers and the tests were able to pass by coincidence.

However, the update of `azcore` from v1.6.0 to v1.7.1 includes a change
that causes the default `retryPolicy` to always deep-copy the request,
and the `retryPolicy` lies between the `pipelineVerifier` and the
`httpHeaderPolicy`. This means the coincidence that was allowing the
tests to pass is now gone - the `pipelineVerifier` and
`httpHeaderPolicy` no longer work with the same underlying `map`.

I removed the header checks after finding this bug because I don't see a
good way to check them.

See
https://learn.microsoft.com/en-us/azure/developer/go/azure-sdk-core-concepts#http-pipeline-flow
for some context on the `azcore` request pipelines.