Refactor hive registration tests
They should no longer error on client creation due to the use of lazy
loading within the client constructor.  This means that during client
creation of a non-typed client, the client will not reach out to the
kube-apiserver to determine which types already exist.  It will only do
so on explicit crud operations against the apiserver.