operator: Add reconcile methods to AROReconciler interface
This introduces two new methods in AROReconciler:

  ReconcileEnabled(ctx, request, cluster) -> (result, error)
  ReconcileDisabled(ctx, request, cluster) -> (result, error)

One or the other method will be called based on the controller's
"enabled" operator flag.

AROController itself now implements the Reconcile method in order
to consistently handle disabled controllers.  Subtypes need only
implement the ReconcileEnabled method, called when the controller
is enabled.

The Reconcile method uses the "template method" design pattern
from object-oriented programming.  This warrants some explanation
since the pattern looks a little odd in Golang.

Some background reading:
https://golangbyexample.com/template-method-design-pattern-golang/

AROController provides ReconcileDisabled but leaves ReconcileEnabled
unimplemented, so ReconcileEnabled is effectively an abstract method.
Therefore AROController itself is not an AROReconciler according to
Golang's interface semantics, but its subtypes are.

Because AROController's Reconcile method calls ReconcileEnabled,
which is abstract, it needs a way to cast itself to an AROReconciler.

This is the purpose of the Reconciler field in AROController.  The
field is set during subtype instantiation:

   r := ExampleReconciler{
       AROController: base.AROController {
           Log:         ...
           Client:      ...
           Name:        ...
           EnabledFlag: ...
       }
   }
   r.Reconciler = r
   return r

The Reconciler field is essentially a virtual method table like
those inherent in polymorphic languages like C++ and Python.  It
allows AROController to call abstract AROReconciler methods that
are implemented by a subtype.