Build and run the container with podman fixes
It respects .dockerignore, thus .bingo directory should be included. When untar an archive of secrets we should use  --no-same-owner to not modify the ownership. No need to check if dev-config.yaml exists when it is already ignored. Furthremore the new build us with no cache, when log in to ACR you will use -ojson flag, and we compare repo tag vs each existing tag