Increase ClusterDeployment readiness timeout in CI
About 95% of recent CI cluster creations have failed on the step that
waits for the Hive ClusterDeployment to be ready. After investigating, I
found that the issue seems to be that it takes a bit longer than 5
minutes for the `Unreachable` status condition on the CD to get into the
desired state. Trying to increase this timeout in CI to see if CI starts
passing consistently. Then once CI passes consistently, maybe we need to
try to figure out why it started taking longer for the CD to be ready.