Makefile: run generators in parallel
I don't know necessarily what the benefits are for the built-in Go
generation stanzas, but `go generate` runs everything without
parallelism, which makes the `make generate` target very slow. We can
hack around this without asking anyone to stop using Go generate markers
by finding packages we need to generate and running them individually.
As long as the output is independent, this is safe.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>