[ARO-15009] fix: add logic to skip subnet update under certain conditions (#4087)
* fix: add logic to skip subnet update under certain conditions
This commit addresses the need for the subnet/write permissions.  Currently, in all scenarios, the
subnet/write permission is used via the subnet.CreateOrUpdate method.  This change adds logic to
determine if the values are already set to Disabled and skips running the subnet.CreateOrUpdate
 method if it is already set appropriately.  This will save the need to make an extra API call and
also allow users to configure things like Azure policy where subnet/write permissions are not needed.

Signed-off-by: Dustin Scott <dscott@redhat.com>

* test: add test for skipping create or update of subnet

Signed-off-by: Dustin Scott <dscott@redhat.com>

* test: add additional test for skip create or update for gateway=true/false

Signed-off-by: Dustin Scott <dscott@redhat.com>

---------

Signed-off-by: Dustin Scott <dscott@redhat.com>