azsecrets: generate typed mocks
There were no previous users of this mock, so it was easy enough to
generate typed mocks. Untyped mocks are insidious for refactors as it's
not possible to see where the mocks need to be updated until the tests
run, which is a shame. We work in a statically typed language for a
reason.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>