Fix Shared RP Documentation (#3812)
* Add missing export word

PARENT_DOMAIN_RESOURCEGROUP was missing an export

* Use single quotes

Stay consistent and all the env vars are set using single quotes instead of double

* Move ADMIN_OBJECT_ID upper

The value of ADMIN_OBJECT_ID is used twice before it was declared as env variable. Now it will be declared beforehand

* Fix indentation leveles, spaces, and numbering

Due to indentation levels the page was rendered wrongly

* Reorganize shared rp doc

Move the non-required procedures at the end under a new seciton and connect the certs creation with the certs rotation

* Use double quotes

Stay consistent and all the env vars are set using double quotes instead of single. When single quotes on the exported variable includes escaped $ then you can create a literal string ''