retry bootstrap steps if an authentication error is returned (#4184)
If there's a delay between when the first party service principal is
granted the necessary roles and the RP tries to perform these actions,
an authorization error is returned. Instead of failing (and returning a
relatively unhelpful error message to the customer), these changes cause
the steps to be retried instead.