cmd: allow loading and comparing RA-style config
A new initiative is moving ARO RP-Config to an Ev2 RA-style, which
allows us to be much more succinct and makes new region build-out
easier. This commit adds the capacity to load the new configuration
file, if one is provided, and to error out if any difference is found.
This behavior is configurable to allow a pipeline to be forced to run
while a diff exists if there's an emergency.

A new command-line tool was added, as well, that diffs a directory of
classic configurations with a directory of new-style ones, to ensure
that there are no changes.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>