fix(prune-tags): actions/checkout set fetch-depth (#4226)
I can't believe it took me this long to remember that actions/checkout
is a shallow checkout by default. That means that only the git objects
directly associated with the given git reference is checked out. All
other branches, tags, and objects are not.

Setting `fetch-depth` to `0` instructions actions/checkout to
effectively carry out a full clone of the repository, including tags
which is what we're interested in for this GitHub Actions Workflow.

Fixes: ARO-14879