.bingo: bump golangci-lint to v2
It's best practice to use the most recent version of the linter with
newer versions of Go, as it's common to have older linters fail to parse
new Go syntax and OOM.

The v2 linter changed how they architect their configuration file, and
provided a `golangci-lint migrate` tool to move previous coinfigurations
over, which was used to do the updates here. No semantic differences are
expected.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>