Makefile: use `gci` formatter for imports
The standard tooling these days for formatting and import order is
`gci`, which is the default formatter in `golangci-lint`. We can remove
our dependency on the legacy OpenShift import formatting tooling by
moving to use `gci`.

I attempted to match the import order we had from the OpenShift tool to
the best of my ability. There are small differences, especially with how
sigs.k8s.io is captured, but it should be immaterial.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>