.golangci.yml: stop excluding directories from linting
Aside from the unfortunate JSON schema deprecation notice, which we
cannot fix, and the importas restrictions we don't want to impose on
generated files, the rest of these are valid linting errors and we
should not be excluding them.

This commit combines the automated `--fix` output with some manual
updates.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>