cmd/aro/deploy: allow prefering RA config (#4244)
If we load and validate the RA config and everything looks good, we can
also start to prefer using that version of the config. Once that change
has time to bake, we can remove the logic for loading more than one
version and simply pass the RA config through the second command-line
argument.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>