Use managed identity credential when running in ev2
This uses a similar implementation as `aro deploy`, which is marked with
a TODO for replacement. Replacement of this implementation should be done
for both commands the same way, as they share the same execution contexts
(both run either in Ev2 for production deployments, or locally/sans-MSI for
localdev deployments).