Makefile: add a target to run the auto-fixer (#4247)
* Makefile: correctly depend on tools

The entire point of having per-tool Makefile targets is so that each
top-level target that requires some tool can ensure that the tools it
uses are installed as necessary. Having a top-level `install-tools`
target is fine, if you need to do it, but using it exclusively loses the
whole point of incremental installs and revalidations from Make.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>

* Makefile: add a target to run the auto-fixer

The --fix argument to the linter will automatically update the AST where
possible to make the codebase conformant.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>

---------

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>