Update aro-dnsmasq-pre.sh.gotmpl
Enable multiple search domains. Update the SEARCHDOMAIN to drop the first column instead of just selecting the first entry. Then modified the searches inject to aro-dns.conf to insert the values in a list format so they will be parsed properly. 

This change enables multiple search terms to be used in ARO